refactor(wallpaper): 拆分来源弹窗的控制区画廊和底部操作
将现有 JSX 和展示属性移入三个类型明确的组件，保留搜索、生成、下载、应用、删除和筛选状态及回调。保持现有 DOM/CSS、文案、渠道和交互，为后续功能分批接入提供边界。
